+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add ocaml-version bound
 - Support OCaml 4.00
 - Bisect is not necessary to run the test suite
+- Add `Record.declare{1,2,3,4}` to build fixed-size layouts
 
 ## v0.1.0